requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.4
django-celery==3.1.17
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.1
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
twilio==9.4.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
